pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "olvid.version.__version__"}

[project]
name = "olvid-bot"
dynamic = ["version"]
authors = [
  { name="Olvid", email="bot@olvid.io" },
]
maintainers = [
  { name="Olvid", email="bot@olvid.io" },
]

keywords = ["olvid", "daemon", "bot"]
description = "An high-level module to create bots in Olvid application"
readme = { file = "README.md" , content-type = "text/markdown"}
license = {file = "LICENSE"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",

    "Intended Audience :: Developers",
    "Topic :: Software Development",

    "Operating System :: OS Independent",
]
dependencies = [
    "grpcio==1.65.4", "grpcio-tools==1.65.4", "protobuf==5.27.3",
    # cli
    "asyncclick==8.1.7.2", "anyio==4.4.0", "uvloop==0.19.0"
]

[project.scripts]
olvid-cli = "olvid:__main__.main"

[project.urls]
GitHug = "https://github.com/olvid-io/Olvid-Bot-Python-Client"
Issues = "https://github.com/olvid-io/Olvid-Bot-Python-Client/issues"
Quickstart = "https://github.com/olvid-io/Olvid-Bot-Quickstart"
Website = "https://olvid.io"